Use props for state and crime in <Explorer />
