+++ package.json
@@ -46,6 +46,7 @@
     "jest": "^17.0.3",
     "js-yaml": "^3.7.0",
     "json-loader": "^0.5.4",
+    "lodash.startcase": "^4.4.0",
     "node-sass": "^3.13.0",
     "nodemon": "^1.11.0",
     "normalize.css": "^5.0.0",
